test: NewCreditCardPageContent 컴포넌트 스토리를 작성하다
